[孙鹏飞]-[修改初始化SDK流程  增加撤回消息之后根据回调处理逻辑  群成员新增 删除不在弹出输入框]

+++ app/src/main/java/top/maxim/im/bmxmanager/ChatManager.java
@@ -108,7 +108,7 @@ public class ChatManager extends BaseManager {
     /**
      * 读取一条消息
      **/
-    public BMXMessage getMessage(int msgId) {
+    public BMXMessage getMessage(long msgId) {
         return mService.getMessage(msgId);
     }
